requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
uWSGI==2.0.20
